requirements.txt
# Cortex - 智能运维系统依赖包

# Web 框架
fastapi==0.104.1
uvicorn[standard]==0.24.0

# 数据库
sqlalchemy==2.0.23
alembic==1.13.0
aiosqlite==0.19.0

# 数据验证
pydantic==2.5.0
pydantic-settings==2.1.0

# LLM SDK
anthropic==0.8.1

# 任务调度
apscheduler==3.10.4

# HTTP 客户端
httpx==0.25.2

# 系统监控
psutil==5.9.6

# 日志
loguru==0.7.2

# Telegram Bot
python-telegram-bot==20.7

# 认证
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
bcrypt==4.1.2
python-multipart==0.0.6
email-validator==2.3.0

# 配置管理
pyyaml==6.0.1

# 测试
pytest==7.4.3
pytest-asyncio==0.21.1
pytest-cov==4.1.0

# 开发工具
black==23.12.1
ruff==0.1.8
mypy==1.7.1
pre-commit==3.5.0
